# web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminWordRequest": {
        "properties": {
          "categoryIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "english": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "german": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "perfectForm": {
            "type": [
              "string",
              "null"
            ]
          },
          "presentForm": {
            "type": [
              "string",
              "null"
            ]
          },
          "preteriteForm": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "english",
          "german"
        ],
        "type": "object"
      },
      "AdminWordView": {
        "properties": {
          "categoryIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "english": {
            "type": "string"
          },
          "german": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "perfectForm": {
            "type": [
              "string",
              "null"
            ]
          },
          "presentForm": {
            "type": [
              "string",
              "null"
            ]
          },
          "preteriteForm": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "CategoryRequest": {
        "properties": {
          "name": {
            "maxLength": 140,
            "minLength": 0,
            "type": "string"
          },
          "slug": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "type": "object"
      },
...
      "CategoryView": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CsvPreview": {
        "properties": {
          "added": {
            "format": "int32",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CsvRowPreview"
            },
            "type": "array"
          },
          "skipped": {
            "format": "int32",
            "type": "integer"
          },
          "updated": {
            "format": "int32",
            "type": "integer"
          },
          "valid": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "CsvRowPreview": {
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "english": {
            "type": "string"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "german": {
            "type": "string"
          },
          "row": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PersonalCategorySummary": {
        "properties": {
          "name": {
            "type": "string"
          },
          "wordCount": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "PersonalWordRequest": {
        "properties": {
          "category": {
            "maxLength": 140,
            "minLength": 0,
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "maxLength": 1000,
            "minLength": 0,
            "type": [
              "string",
              "null"
            ]
          },
          "english": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "german": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "english",
          "german"
        ],
        "type": "object"
      },
      "PersonalWordView": {
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "english": {
            "type": "string"
          },
          "german": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "OIDC access token issued for the WörterSee API.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "WörterSee Cloud"
    },
    "description": "Versioned contracts for vocabulary, learning and account data.",
    "title": "WörterSee Cloud API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/categories": {
      "get": {
        "operationId": "categories",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin-vocabulary-controller"
        ]
      },
      "post": {
        "operationId": "createCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryView"
                }
              }
            },
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin-vocabulary-controller"
        ]
      }
    },
...
    "/api/v1/admin/words/{id}": {
      "delete": {
        "operationId": "deleteWord",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin-vocabulary-controller"
        ]
      },
      "put": {
        "operationId": "updateWord",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminWordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminWordView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "admin-vocabulary-controller"
        ]
      }
    },
...
    "/api/v1/personal-words": {
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonalWordView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "personal-word-controller"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalWordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalWordView"
                }
              }
            },
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "personal-word-controller"
        ]
      }
    },
    "/api/v1/personal-words/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "personal-word-controller"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalWordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalWordView"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "personal-word-controller"
        ]
      }
    },
    "/api/v1/personal-words/categories": {
      "get": {
        "operationId": "categories_2",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonalCategorySummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "personal-word-controller"
        ]
      }
    },
    "/api/v1/personal-words/csv/export": {
      "get": {
        "operationId": "export",
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "personal-word-controller"
        ]
      }
    },
    "/api/v1/personal-words/csv/import": {
      "post": {
        "operationId": "importCsv",
        "requestBody": {
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CsvPreview"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "personal-word-controller"
        ]
      }
    },
    "/api/v1/personal-words/csv/preview": {
      "post": {
        "operationId": "preview",
        "requestBody": {
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CsvPreview"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "personal-word-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ]
}